Refactored code for client options supplied logger
