+++ src/IO.Ably.Shared/InternalLogger.cs
@@ -4,6 +4,44 @@ using System.Text;
 
 namespace IO.Ably
 {
+    /// <summary>Level of a log message.</summary>
+    public enum LogLevel : byte
+    {
+        /// <summary>
+        /// Verbose setting. Logs everything.
+        /// </summary>
+        Debug = 0,
+
+        /// <summary>
+        /// Warning setting. Logs clues that something is not 100% right.
+        /// </summary>
+        Warning,
+
+        /// <summary>
+        /// Error setting. Logs errors
+        /// </summary>
+        Error,
+
+        /// <summary>
+        /// None setting. No logs produced
+        /// </summary>
+        None = 99
+    }
+
+    /// <summary>The default logger implementation, that writes to debug output.</summary>
+    internal class DefaultLoggerSink : ILoggerSink
+    {
+        private readonly object _syncRoot = new object();
+
+        public void LogEvent(LogLevel level, string message)
+        {
+            lock (_syncRoot)
+            {
+                Debug.WriteLine($"Ably: [{level}] {message}");
+            }
+        }
+    }
+
     internal class InternalLogger : IInternalLogger
     {
         private InternalLogger()
