+++ src/IO.Ably.Tests.Shared/Extensions/MessageExtensions.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Text;
+using IO.Ably.Tests.Shared.Helpers;
 
 namespace IO.Ably.Tests.Shared.Extensions
 {
