+++ src/IO.Ably.Tests.Shared/IO.Ably.Tests.Shared.projitems
@@ -31,6 +31,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)EncryptionSpecs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ErrorInfoTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\ChannelOptionsExtensions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Helpers\DefaultLogger.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\MessageExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\StringExtensionsTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Helpers\JAssert.cs" />
