+++ src/IO.Ably.Tests.Shared/MessageEncodes/JsonEncoderTests.cs
@@ -1,5 +1,6 @@
 using FluentAssertions;
 using IO.Ably.MessageEncoders;
+using IO.Ably.Tests.Shared.Helpers;
 using Newtonsoft.Json.Linq;
 using Xunit;
